Expose I/O operations counters per tablespace.
Although InnoDB exposes counters for the number of pages and bytes
read/written, it is not possible to break down these numbers per table,
which would help in determining the I/O distribution among tables (and
the tablespaces they belong to).

This change introduces a set of per tablespace counters for read, write
and flush operations. The counters are incremented whenever data is read,
written or flushed to an InnoDB space. Additionally, the number of bytes
read and written are also counted. These statistics are kept in the
tablespace memory cache and, hence, are ephemeral. In the unlikely case
a tablespace is removed from the cache, the counters values are lost.

A new information schema table named INNODB_SPACE_STATS is introduced
to expose these statistics counters. Each row of the table represents
a space in the tablespace memory cache, and each column represents a
counter value.